some cleanup in TCPSlaveBase: - move protected variables into private class - merge constructors - fix const'ness - make port a quint16 instead of a signed integer - indentation and similar fixups
svn path=/trunk/KDE/kdelibs/; revision=657250

+++ kio/kio/dataslave.cpp
@@ -145,7 +145,7 @@ bool DataSlave::suspended() {
   return _suspended;
 }
 
-void DataSlave::setHost(const QString &/*host*/, int /*port*/,
+void DataSlave::setHost(const QString &/*host*/, quint16 /*port*/,
                      const QString &/*user*/, const QString &/*passwd*/) {
   // irrelevant -> will be ignored
 }
